pyproject.toml
[project]
name = "properties-analytics"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "numpy (>=2.2.6,<3.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "pandas-datareader (>=0.10.0,<0.11.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
jupyterlab = "^4.4.3"
notebook = "^7.4.3"
ipykernel = "^6.29.5"
black = "^25.1.0"
isort = "^6.0.1"
pandas-datareader = "^0.10.0"
python-dotenv = "^1.1.0"

